ajax

+++ SkyWebApplication/Controllers/CategoryController.cs
@@ -117,42 +117,59 @@ namespace SkyWebApplication.Controllers
         }
 
         // GET: /Category/Create
-        public ActionResult Create()
-        {
-            return View();
-        }
+        //public ActionResult Create()
+        //{
+        //    return View();
+        //}
 
         // POST: /Category/Create
         // 为了防止“过多发布”攻击，请启用要绑定到的特定属性，有关 
         // 详细信息，请参阅 http://go.microsoft.com/fwlink/?LinkId=317598。
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include="ID,CategoryName,CategoryInfo,CategoryParentID,CategoryStatus,CategorySort")] Category category)
+        public ActionResult Create([Bind(Include = "CategoryName,CategoryInfo,CategoryParentID,CategoryStatus,CategorySort")] Category category)
         {
+
             if (ModelState.IsValid)
             {
                 db.Categorys.Add(category);
                 db.SaveChanges();
-                return RedirectToAction("Index");
-            }
 
-            return View(category);
+                return RedirectToAction("Dictionary", "Home");
+            }
+            return RedirectToAction("Dictionary", "Home");
         }
+
+
+        //[HttpPost]
+        //[ValidateAntiForgeryToken]
+        //public ActionResult Create([Bind(Include = "ID,CategoryName,CategoryInfo,CategoryParentID,CategoryStatus,CategorySort")] Category category)
+        //{
+        //    if (ModelState.IsValid)
+        //    {
+        //        db.Categorys.Add(category);
+        //        db.SaveChanges();
+        //        return RedirectToAction("Dictionary", "Home");
+        //    }
+
+        //    return RedirectToAction("Dictionary", "Home");
+        //}
+
 
         // GET: /Category/Edit/5
-        public ActionResult Edit(int? id)
-        {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Category category = db.Categorys.Find(id);
-            if (category == null)
-            {
-                return HttpNotFound();
-            }
-            return View(category);
-        }
+        //public ActionResult Edit(int? id)
+        //{
+        //    if (id == null)
+        //    {
+        //        return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+        //    }
+        //    Category category = db.Categorys.Find(id);
+        //    if (category == null)
+        //    {
+        //        return HttpNotFound();
+        //    }
+        //    return View(category);
+        //}
 
         // POST: /Category/Edit/5
         // 为了防止“过多发布”攻击，请启用要绑定到的特定属性，有关 
@@ -165,37 +182,72 @@ namespace SkyWebApplication.Controllers
             {
                 db.Entry(category).State = EntityState.Modified;
                 db.SaveChanges();
-                return RedirectToAction("Index");
+                return RedirectToAction("Dictionary", "Home");
             }
-            return View(category);
+            return RedirectToAction("Dictionary", "Home");
         }
 
         // GET: /Category/Delete/5
-        public ActionResult Delete(int? id)
-        {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Category category = db.Categorys.Find(id);
-            if (category == null)
-            {
-                return HttpNotFound();
-            }
-            return View(category);
-        }
+        //public ActionResult Delete(int? id)
+        //{
+        //    if (id == null)
+        //    {
+        //        return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+        //    }
+        //    Category category = db.Categorys.Find(id);
+        //    if (category == null)
+        //    {
+        //        return HttpNotFound();
+        //    }
+        //    return View(category);
+        //}
 
         // POST: /Category/Delete/5
-        [HttpPost, ActionName("Delete")]
+        //[HttpPost, ActionName("Delete")]
+        //[ValidateAntiForgeryToken]
+        //public ActionResult DeleteConfirmed(int id)
+        //{
+        //    Category category = db.Categorys.Find(id);
+        //    db.Categorys.Remove(category);
+        //    db.SaveChanges();
+        //    return RedirectToAction("Index");
+        //}
+        [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public JsonResult Delete(int? id)
         {
-            Category category = db.Categorys.Find(id);
-            db.Categorys.Remove(category);
-            db.SaveChanges();
-            return RedirectToAction("Index");
-        }
+            Message msg = new Message();
 
+            if (id == null)
+            {
+                msg.MessageStatus = "false";
+                msg.MessageInfo = "找不到ID";
+            }
+            else
+            {
+                var childcategorys = from s in db.Categorys
+                                     orderby s.CategorySort ascending
+                                     where s.CategoryParentID == id
+                                     select s;
+
+                if (childcategorys.Count() > 0)
+                {
+                    msg.MessageStatus = "false";
+                    msg.MessageInfo = "此ID节点下有子节点，不能删除";
+
+                }
+                else
+                {
+                    Category category = db.Categorys.Find(id);
+                    db.Categorys.Remove(category);
+                    db.SaveChanges();
+                    msg.MessageStatus = "true";
+                    msg.MessageInfo = "删除成功";                
+                }
+            }
+
+            return Json(msg, JsonRequestBehavior.AllowGet);
+        }
         protected override void Dispose(bool disposing)
         {
             if (disposing)
